v2_code/requirements.txt
flask
werkzeug
requests
sqlalchemy
Flask-SQLAlchemy
flask-login
pandas
flask_restful
flask-security
weasyprint
pyjwt
redis
celery
celery[redis]
jinja
